refactor(dict): Make room for trie logic
